Refactor servlet packages
