+++ rm-portal/src/main/java/org/ow2/proactive_grid_cloud_portal/rm/server/nodesource/serialization/export/ExportToFileServlet.java
@@ -23,7 +23,7 @@
  * If needed, contact us to obtain a release under GPL Version 2 or 3
  * or a different license than the AGPL.
  */
-package org.ow2.proactive_grid_cloud_portal.rm.server.serialization;
+package org.ow2.proactive_grid_cloud_portal.rm.server.nodesource.serialization.export;
 
 import java.io.IOException;
 
@@ -45,9 +45,9 @@ import org.slf4j.LoggerFactory;
  * content is expected to be in JSON format, containing the configuration of a
  * node source.
  */
-public class ExportNodeSourceToFileServlet extends HttpServlet {
+public class ExportToFileServlet extends HttpServlet {
 
-    private static final Logger LOGGER = LoggerFactory.getLogger(ExportNodeSourceToFileServlet.class);
+    private static final Logger LOGGER = LoggerFactory.getLogger(ExportToFileServlet.class);
 
     private static final String NODE_SOURCE_FILE_NAME_SUFFIX = "-configuration.json";
